moved updating of innovation flows to specify a template ID; added some docs; added some more renames
