+++ graphql-samples/mutations/create/create-opportunity
@@ -10,9 +10,9 @@ query variables:
   "opportunityData":
   {
     "challengeID": "uuid",
+    "innovationFlowTemplateID": "uuid",
     "displayName": "Balance the grid",
     "nameID": "balance-grid",
-    "lifecycleTemplate": "EXTENDED",
     "context": {
         "tagline": "How might we incentivize consumers to communicate energy demand and production to allow all stakeholders to balance the grid?",
         "background": "Our power system is becoming increasingly more decentralized and complex. By using solar PV, consumers have become electricity producers as well, and the number of these ‘prosumers’ is increasing. All the while, the electricity demand is rising due to the electrification of transport and industry. Cities around the world are electrifying their public transport systems, but if all busses would be charged around the time people are in the kitchen preparing dinner, the current grid would not be able to cope. The energy sector is striving to make the best of these developments, but so far the sector remains fragmented with a large number of initiatives and pilots spread over start-ups, scale-ups, countries, DSOs, TSOs, utilities, and regulators. Balancing the power grid is and will be more and more difficult and expensive. While technological advances have already disrupted many areas of the power system and, for example, empowered consumers to become producers and traders of their own electricity, the balancing markets still resemble an exclusive club of the big industry players. How can we enable also small consumers and prosumers to contribute to balancing? ",
