+++ graphql-samples/mutations/update/update-challenge-innovation-flow
@@ -1,0 +1,14 @@
+mutation updateChallengeInnovationFlow($challengeData: UpdateChallengeInnovationFlowInput!) {
+  updateChallengeInnovationFlow(challengeData: $challengeData) {
+    id
+  }
+}
+
+
+query variables:
+{
+  "challengeData": {
+    "challengeID": "uuid",
+    "innovationFlowTemplateID": "uuid"
+  }
+}
