--- graphql-samples/mutations/update/update-opportunity-lifecycle
@@ -1,15 +1,0 @@
-mutation updateOpportunityLifecycle($opportunityData: UpdateOpportunityLifecycleInput!) {
-  updateOpportunityLifecycle(opportunityData: $opportunityData) {
-    id
-    value
-  }
-}
-
-
-query variables:
-{
-  "opportunityData": {
-    "opportunityID": "uuid",
-    "lifecycleDefinition": "{\"id\":\"opportunity-lifecycle-default-4\",\"context\":{\"parentID\":\"2f2f3241-a70c-4887-a2c8-262059f36dfc\"},\"initial\":\"new\",\"states\":{\"new\":{\"on\":{\"REFINE\":{\"target\":\"beingRefined\",\"cond\":\"challengeStateUpdateAuthorized\"},\"ABANDONED\":{\"target\":\"abandoned\",\"cond\":\"challengeStateUpdateAuthorized\"}}},\"beingRefined\":{\"on\":{\"ACTIVE\":{\"target\":\"inProgress\",\"cond\":\"challengeStateUpdateAuthorized\"},\"ABANDONED\":{\"target\":\"abandoned\",\"cond\":\"challengeStateUpdateAuthorized\"}}},\"inProgress\":{\"entry\":[\"sampleEvent\"],\"on\":{\"COMPLETED\":{\"target\":\"complete\",\"cond\":\"challengeStateUpdateAuthorized\"},\"ABANDONED\":{\"target\":\"abandoned\",\"cond\":\"challengeStateUpdateAuthorized\"}}},\"complete\":{\"on\":{\"ARCHIVE\":\"archived\",\"ABANDONED\":\"abandoned\"}},\"abandoned\":{\"on\":{\"REOPEN\":\"inProgress\",\"ARCHIVE\":\"archived\"}},\"archived\":{\"type\":\"final\"}}}"
-  }
-}
