+++ src/domain/challenge/challenge/challenge.service.ts
@@ -50,7 +50,7 @@ import { PreferenceType } from '@common/enums/preference.type';
 import { CredentialDefinition } from '@domain/agent/credential/credential.definition';
 import { CommunityRole } from '@common/enums/community.role';
 import { challengeCommunityPolicy } from './challenge.community.policy';
-import { UpdateChallengeInnovationFlowInput } from './dto/challenge.dto.update.lifecycle';
+import { UpdateChallengeInnovationFlowInput } from './dto/challenge.dto.update.innovation.flow';
 import { ICollaboration } from '@domain/collaboration/collaboration/collaboration.interface';
 import { LifecycleTemplateService } from '@domain/template/lifecycle-template/lifecycle.template.service';
 import { LifecycleType } from '@common/enums/lifecycle.type';
@@ -190,7 +190,14 @@ export class ChallengeService {
       );
     }
 
-    challenge.lifecycle.machineDef = challengeData.innovationFlowDefinition;
+    const machineConfig: ILifecycleDefinition =
+      await this.lifecycleTemplateService.getLifecycleDefinitionFromTemplate(
+        challengeData.innovationFlowTemplateID,
+        challenge.hubID,
+        LifecycleType.CHALLENGE
+      );
+
+    challenge.lifecycle.machineDef = JSON.stringify(machineConfig);
     challenge.lifecycle.machineState = '';
 
     return await this.challengeRepository.save(challenge);
