+++ src/domain/challenge/challenge/dto/challenge.dto.update.innovation.flow.ts
@@ -1,6 +1,5 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { UUID } from '@domain/common/scalars';
-import { LifecycleDefinitionScalar } from '@domain/common/scalars/scalar.lifecycle.definition';
 
 @InputType()
 export class UpdateChallengeInnovationFlowInput {
@@ -9,9 +8,9 @@ export class UpdateChallengeInnovationFlowInput {
   })
   challengeID!: string;
 
-  @Field(() => LifecycleDefinitionScalar, {
+  @Field(() => UUID, {
     nullable: false,
-    description: 'The Innovation Flow Definition to use for this Challenge.',
+    description: 'The Innovation Flow template to use for the Challenge.',
   })
-  innovationFlowDefinition!: string;
+  innovationFlowTemplateID!: string;
 }
