+++ src/domain/collaboration/opportunity/dto/opportunity.dto.create.ts
@@ -7,6 +7,9 @@ export class CreateOpportunityInput extends CreateBaseChallengeInput {
   @Field(() => UUID, { nullable: false })
   challengeID!: string;
 
-  @Field(() => UUID, { nullable: false })
+  @Field(() => UUID, {
+    nullable: false,
+    description: 'The Innovation Flow template to use for the Opportunity.',
+  })
   innovationFlowTemplateID!: string;
 }
