+++ src/domain/collaboration/opportunity/dto/opportunity.dto.update.innovation.flow.ts
@@ -1,6 +1,5 @@
 import { Field, InputType } from '@nestjs/graphql';
 import { UUID } from '@domain/common/scalars';
-import { LifecycleDefinitionScalar } from '@domain/common/scalars/scalar.lifecycle.definition';
 
 @InputType()
 export class UpdateOpportunityInnovationFlowInput {
@@ -9,9 +8,9 @@ export class UpdateOpportunityInnovationFlowInput {
   })
   opportunityID!: string;
 
-  @Field(() => LifecycleDefinitionScalar, {
+  @Field(() => UUID, {
     nullable: false,
-    description: 'The Innovation Flow Definition to use for this Opportunity.',
+    description: 'The Innovation Flow template to use for the Opportunity.',
   })
-  innovationFlowDefinition!: string;
+  innovationFlowTemplateID!: string;
 }
