+++ src/domain/common/authorization-policy/authorization.policy.service.ts
@@ -312,7 +312,7 @@ export class AuthorizationPolicyService {
         AuthorizationPrivilege.GRANT,
         AuthorizationPrivilege.READ,
         AuthorizationPrivilege.UPDATE,
-        AuthorizationPrivilege.UPDATE_LIFECYCLE,
+        AuthorizationPrivilege.UPDATE_INNOVATION_FLOW,
         AuthorizationPrivilege.DELETE,
       ],
       AuthorizationCredential.GLOBAL_ADMIN
